add_retrofit_service_interface: Added the Retrofit interface class and the temporary sample call to the network for temporary display in main activity.

+++ src/main/java/com/arleneg/android/scrumptiousbakes/data/Ingredient.java
@@ -12,7 +12,7 @@ public class Ingredient implements Parcelable
 
     @SerializedName("quantity")
     @Expose
-    private long quantity;
+    private double quantity;
 
     @SerializedName("measure")
     @Expose
@@ -38,7 +38,7 @@ public class Ingredient implements Parcelable
     };
 
     protected Ingredient(Parcel in) {
-        this.quantity = ((long) in.readValue((long.class.getClassLoader())));
+        this.quantity = ((double) in.readValue((double.class.getClassLoader())));
         this.measure = ((String) in.readValue((String.class.getClassLoader())));
         this.ingredient = ((String) in.readValue((String.class.getClassLoader())));
     }
@@ -56,18 +56,18 @@ public class Ingredient implements Parcelable
      * @param ingredient
      * @param quantity
      */
-    public Ingredient(long quantity, String measure, String ingredient) {
+    public Ingredient(double quantity, String measure, String ingredient) {
         super();
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
-    public long getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(long quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
